+++ maya_tools/modeling_tools/shaders.py
@@ -1,38 +1,4 @@
 import maya.cmds as cmds
-import os.path
-import constants
-
-def build_structure(grp_name=None):
-    """ Build asset structure based on file name if a scene_name is not given
-    Args:
-        grp_name(str): name of the parent group, defaults to the file name if None
-    """
-    # if package name is not defined, will name it based on the file name
-    if not grp_name:
-        grp_name = os.path.basename(cmds.file(q=True, sn=True))
-
-    # desired list of groups we want to create
-    asset_groups = constants.ASSET_GROUPS
-
-    # creates the groups and names them
-    for asset_group in asset_groups:
-        cmds.CreateEmptyGroup()
-        cmds.rename(asset_group)
-
-    # groups our created groups and names the group based on the scene
-    cmds.group(grp_name[:-3], asset_groups)
-
-def export_open():
-    file_path = cmds.file(sceneName=True, query=True)
-    file_name = os.path.basename(file_path)[:-3]
-
-    if not cmds.objExists(file_name):
-        print("Nothing to export, no object named : {}".format(file_name))
-    else:
-        cmds.select(file_name)
-
-    cmds.file(file_path, type="mayaAscii", exportSelected=True, force=True)
-    cmds.file(file_path, type="mayaAscii", open=True, force=True)
 
 def compare_shaders():
     shader_groups = cmds.ls(type="shadingEngine")
